fix: suspense boundaries on payment success
